+++ src/admin/dashboard/static-data.js
@@ -1,0 +1,37 @@
+export const staticData = [
+  {
+    icon: 'people',
+    title: 'Utilizatori',
+    statistics: '71.897',
+    link: '#',
+    linkText: 'Detalii',
+  },
+  {
+    icon: 'envelope',
+    title: 'Conturi Active',
+    statistics: '62.175',
+    link: '#',
+    linkText: 'Detalii',
+  },
+  {
+    icon: 'pointer',
+    title: 'Număr de lecții',
+    statistics: '415',
+    link: '#',
+    linkText: 'Detalii',
+  },
+  {
+    icon: 'people',
+    title: 'Lectii cu peste 1.000 de accesari',
+    statistics: '210',
+    link: '#',
+    linkText: 'Detalii',
+  },
+  {
+    icon: 'envelope',
+    title: 'Lectii cu sub 100 de accesari',
+    statistics: '15',
+    link: '#',
+    linkText: 'Detalii',
+  },
+]
