+++ src/admin/dashboard/title/index.jsx
@@ -1,11 +1,10 @@
 import React from 'react'
 import * as Material from '@material-ui/core'
 
-export const Title = () => {
-    return(
-    <Material.Box display={'inline-block'}>
-        <Material.Typography variant={'h4'}>
-            <strong>Dashboard</strong>
-        </Material.Typography>
-    </Material.Box>
-)}
+export const Title = () => (
+  <Material.Box display={'inline-block'}>
+    <Material.Typography variant={'h4'}>
+      <strong>Dashboard</strong>
+    </Material.Typography>
+  </Material.Box>
+)
